Move Environment::isLocalBranchEqualTo to CommitishComparator::isLocalBranchEqualTo
