+++ src/main/php/CommandLine/Library/Environment.php
@@ -14,20 +14,16 @@ use function Safe\realpath;
  */
 class Environment
 {
-    private ?string $localHeadHash = null;
     private ProcessRunner $processRunner;
-    private GitInputValidator $gitInputValidator;
     private EnhancedFileInfoFactory $enhancedFileInfoFactory;
     /** @var string */
     private const GIT = 'git';
 
     public function __construct(
         ProcessRunner $processRunner,
-        GitInputValidator $gitInputValidator,
         EnhancedFileInfoFactory $enhancedFileInfoFactory
     ) {
         $this->processRunner = $processRunner;
-        $this->gitInputValidator = $gitInputValidator;
         $this->enhancedFileInfoFactory = $enhancedFileInfoFactory;
     }
 
@@ -61,30 +57,5 @@ class Environment
         $packagePath = dirname(__DIR__, 5);
         $enhancedFileInfo = $this->enhancedFileInfoFactory->buildFromPath($packagePath);
         return $enhancedFileInfo;
-    }
-
-    /**
-     * Compare if the HEAD of $target Branch equals the HEAD of the local branch.
-     */
-    public function isLocalBranchEqualTo(?string $targetBranch): bool
-    {
-        if (!$this->gitInputValidator->isCommitishValid($targetBranch)) {
-            return false;
-        }
-        if ($this->localHeadHash === null) {
-            $this->localHeadHash = $this->commitishToHash('HEAD');
-        }
-
-        $targetCommitHash = $this->commitishToHash($targetBranch);
-
-        return $targetCommitHash === $this->localHeadHash;
-    }
-
-    /**
-     * Converts a commit-tish to a commit hash.
-     */
-    private function commitishToHash(?string $branchName): string
-    {
-        return $this->processRunner->runAsProcess(self::GIT, 'rev-list', '-n 1', $branchName);
     }
 }
