+++ tests/Unit/CommandLine/Library/EnvironmentTest.php
@@ -9,7 +9,6 @@ use Mockery\MockInterface;
 use PHPUnit\Framework\TestCase;
 use Zooroyal\CodingStandard\CommandLine\Factories\EnhancedFileInfoFactory;
 use Zooroyal\CodingStandard\CommandLine\Library\Environment;
-use Zooroyal\CodingStandard\CommandLine\Library\GitInputValidator;
 use Zooroyal\CodingStandard\CommandLine\Library\ProcessRunner;
 use Zooroyal\CodingStandard\CommandLine\ValueObjects\EnhancedFileInfo;
 use Zooroyal\CodingStandard\Tests\Tools\SubjectFactory;
@@ -85,108 +84,6 @@ class EnvironmentTest extends TestCase
         $result = $this->subject->getPackageDirectory();
 
         self::assertSame($this->mockedEnhancedFileInfo, $result);
-    }
-
-    /**
-     * @test
-     */
-    public function isLocalBranchEqualToReturnsTrueIfCommitHashesAreEqual(): void
-    {
-        $mockedBranchName = 'my/mocked/branch';
-        $mockedCommitHash = '123qwe0';
-
-        $this->subjectParameters[GitInputValidator::class]->shouldReceive('isCommitishValid')->once()
-            ->with($mockedBranchName)->andReturn(true);
-
-        $this->subjectParameters[ProcessRunner::class]->shouldReceive('runAsProcess')->once()
-            ->with('git', 'rev-list', '-n 1', 'HEAD')->andReturn($mockedCommitHash);
-        $this->subjectParameters[ProcessRunner::class]->shouldReceive('runAsProcess')->once()
-            ->with('git', 'rev-list', '-n 1', $mockedBranchName)->andReturn($mockedCommitHash);
-
-        $result = $this->subject->isLocalBranchEqualTo($mockedBranchName);
-        self::assertTrue($result);
-    }
-
-    /**
-     * @test
-     */
-    public function isLocalBranchEqualToReturnsFalseIfCommitHashesAreUnequal(): void
-    {
-        $mockedBranchName = 'my/mocked/branch';
-        $mockedCommitHash = '123qwe0';
-        $mockedLocalCommitHash = '0ewq321';
-
-        $this->subjectParameters[GitInputValidator::class]->shouldReceive('isCommitishValid')->once()
-            ->with($mockedBranchName)->andReturn(true);
-
-        $this->subjectParameters[ProcessRunner::class]->shouldReceive('runAsProcess')->once()
-            ->with('git', 'rev-list', '-n 1', 'HEAD')->andReturn($mockedLocalCommitHash);
-        $this->subjectParameters[ProcessRunner::class]->shouldReceive('runAsProcess')->once()
-            ->with('git', 'rev-list', '-n 1', $mockedBranchName)->andReturn($mockedCommitHash);
-
-        $result = $this->subject->isLocalBranchEqualTo($mockedBranchName);
-        self::assertFalse($result);
-    }
-
-    /**
-     * @test
-     */
-    public function isLocalBranchEqualToCachesLocalHeadHash(): void
-    {
-        $mockedBranchName = 'my/mocked/branch';
-        $mockedCommitHash = '123qwe0';
-        $mockedLocalCommitHash = '0ewq321';
-
-        $this->subjectParameters[GitInputValidator::class]->shouldReceive('isCommitishValid')->twice()
-            ->with($mockedBranchName)->andReturn(true);
-
-        $this->subjectParameters[ProcessRunner::class]->shouldReceive('runAsProcess')->once()
-            ->with('git', 'rev-list', '-n 1', 'HEAD')->andReturn($mockedLocalCommitHash);
-        $this->subjectParameters[ProcessRunner::class]->shouldReceive('runAsProcess')->twice()
-            ->with('git', 'rev-list', '-n 1', $mockedBranchName)->andReturn($mockedCommitHash);
-
-        $this->subject->isLocalBranchEqualTo($mockedBranchName);
-        $result = $this->subject->isLocalBranchEqualTo($mockedBranchName);
-        self::assertFalse($result);
-    }
-
-    /**
-     * @test
-     */
-    public function isLocalBranchEqualToReturnsFalseIfParameterNoBranch(): void
-    {
-        $mockedBranchName = 'my/mocked/branch';
-        $mockedLocalCommitHash = '0ewq321';
-
-        $this->subjectParameters[GitInputValidator::class]->shouldReceive('isCommitishValid')->once()
-            ->with($mockedBranchName)->andReturn(false);
-
-        $this->subjectParameters[ProcessRunner::class]->shouldReceive('runAsProcess')->never()
-            ->with('git', 'rev-list', '-n 1', 'HEAD')->andReturn($mockedLocalCommitHash);
-        $this->subjectParameters[ProcessRunner::class]->shouldReceive('runAsProcess')->never()
-            ->with('git', 'rev-list', '-n 1', $mockedBranchName);
-
-        $result = $this->subject->isLocalBranchEqualTo($mockedBranchName);
-        self::assertFalse($result);
-    }
-
-    /**
-     * @test
-     */
-    public function isLocalBranchEqualToWithNull(): void
-    {
-        $mockedBranchName = null;
-
-        $this->subjectParameters[GitInputValidator::class]->shouldReceive('isCommitishValid')->once()
-            ->with($mockedBranchName)->andReturn(false);
-
-        $this->subjectParameters[ProcessRunner::class]->shouldReceive('runAsProcess')->never()
-            ->with('git', 'rev-list', '-n 1', 'HEAD');
-        $this->subjectParameters[ProcessRunner::class]->shouldReceive('runAsProcess')->never()
-            ->with('git', 'rev-list', '-n 1', $mockedBranchName);
-
-        $result = $this->subject->isLocalBranchEqualTo($mockedBranchName);
-        self::assertFalse($result);
     }
 
 }
